Rect.h: fix signedness warnings

+++ ttbasic/Rect.h
@@ -76,7 +76,7 @@ public:
 		if (r.width == 0 || r.height == 0)
 			return true;
 
-		for(size_t i = 0; i < rects.size(); ++i)
+		for(int i = 0; i < rects.size(); ++i)
 			if (!Disjoint(rects[i], r))
 				return false;
 		return true;
